Servicios funcionales todo OK

+++ src/main/java/com/uniminuto/clinica/repository/MedicoRepository.java
@@ -9,6 +9,6 @@ import org.springframework.stereotype.Repository;
  * @author lmora
  */
 @Repository
-public interface MedicoRepository extends JpaRepository<Medico, Integer> {
+public interface MedicoRepository extends JpaRepository<Medico, Long> {
     
 }
